Extract SleeperInstanceDriver
